+++ lib/src/lexemes/and.dart
@@ -1,4 +1,5 @@
-import 'package:token_parser/src/internal/extension.dart';
+import 'package:token_parser/src/debug.dart';
+import 'package:token_parser/src/extensions/tokenize.dart';
 import 'package:token_parser/src/token.dart';
 import 'package:token_parser/src/lexemes/abstract/parent.dart';
 import 'package:token_parser/src/lexemes/abstract/bound.dart';