+++ lib/src/lexemes/character.dart
@@ -1,5 +1,7 @@
+import 'package:token_parser/src/debug.dart';
 import 'package:token_parser/src/error.dart';
-import 'package:token_parser/src/internal/extension.dart';
+import 'package:token_parser/src/extension.dart';
+import 'package:token_parser/src/extensions/tokenize.dart';
 import 'package:token_parser/src/lexemes/pattern.dart';
 import 'package:token_parser/src/token.dart';